refactor: extract checks modal to self contained component
This is to help with re-usability
